+++ src/controllers/modal/interface.js
@@ -1,14 +1,22 @@
-import { ZalgoPromise } from 'zalgo-promise';
+import arrayFind from 'core-js-pure/stable/array/find';
+import { ZalgoPromise } from 'zalgo-promise/src';
 
-import { logger, memoizeOnProps, getCurrentTime, viewportHijack, awaitWindowLoad } from '../../utils';
+import {
+    logger,
+    memoizeOnProps,
+    getCurrentTime,
+    awaitWindowLoad,
+    getInlineOptions,
+    isElement,
+    globalState,
+    objectMerge,
+    getProductForOffer
+} from '../../utils';
 import { Modal } from '../../zoid/modal';
 
-export default memoizeOnProps(
-    ({ account, merchantId, currency, amount, buyerCountry, offer, onReady, onCalculate, onApply, onClose, index }) => {
-        const [hijackViewport] = viewportHijack();
-
-        const { render, hide, updateProps, state } = Modal({
-            index,
+const memoizedModal = memoizeOnProps(
+    ({ account, merchantId, currency, amount, buyerCountry, offer, onReady, onCalculate, onApply, onClose }) => {
+        const { render, hide, updateProps, state, event } = Modal({
             account,
             merchantId,
             currency,
@@ -20,6 +28,8 @@ export default memoizeOnProps(
             onApply,
             onClose
         });
+        // Fired from inside the default onReady callback
+        const modalReady = new ZalgoPromise(resolve => event.once('ready', resolve));
 
         let renderProm;
         const renderModal = (selector = 'body') => {
@@ -29,30 +39,37 @@ export default memoizeOnProps(
                 renderProm = awaitWindowLoad
                     // Give priority to other merchant scripts waiting for the load event
                     .then(() => ZalgoPromise.delay(0))
-                    .then(() => render(selector))
-                    // The render promise will resolve before Preact renders and picks up changes
-                    // via updateProps so a small delay is added after the initial "render" promise
-                    .then(() => ZalgoPromise.delay(100));
+                    .then(() => ZalgoPromise.all([render(selector), modalReady]));
                 hide();
             }
 
             return renderProm;
         };
 
-        const showModal = (newOptions = {}) => {
+        const showModal = (options = {}) => {
+            const newOptions = isElement(options) ? getInlineOptions(options) : options;
+
             state.renderStart = getCurrentTime();
+
             if (!renderProm) {
                 renderProm = renderModal('body');
             }
-            return renderProm.then(() => {
-                hijackViewport();
 
-                logger.track({
-                    index: newOptions.index,
-                    et: 'CLIENT_IMPRESSION',
-                    event_type: 'modal-open'
+            const requestedProduct = getProductForOffer(options.offer);
+
+            if (
+                typeof options.offer !== 'undefined' &&
+                Array.isArray(state.products) &&
+                !arrayFind(state.products, supportedProduct => supportedProduct === requestedProduct)
+            ) {
+                logger.warn('invalid_option_value', {
+                    location: 'offer',
+                    description: `Expected one of ["${state.products.join('", "')}"] but received "${options.offer}".`
                 });
+                return ZalgoPromise.resolve();
+            }
 
+            return renderProm.then(() => {
                 return updateProps({
                     visible: true,
                     ...newOptions
@@ -64,8 +81,14 @@ export default memoizeOnProps(
             if (!renderProm) {
                 renderProm = renderModal('body');
             }
+
             return renderProm.then(() => updateProps({ visible: false }));
         };
+
+        // Expose these functions through the computed onReady callback
+        // for merchant integrations
+        state.show = showModal;
+        state.hide = hideModal;
 
         // Follow existing zoid interface
         return {
@@ -77,3 +100,5 @@ export default memoizeOnProps(
     },
     ['account', 'merchantId']
 );
+
+export default options => memoizedModal(objectMerge(globalState.config, options));