+++ tests/unit/spec/src/components/modal/lib/providers/transition.test.js
@@ -8,7 +8,8 @@ const defaultXProps = {
     show: jest.fn().mockResolvedValue(null),
     hide: jest.fn().mockResolvedValue(null),
     onProps: jest.fn(),
-    onClose: jest.fn()
+    onClose: jest.fn(),
+    onShow: jest.fn()
 };
 
 jest.mock('src/components/lib/providers/xprops');
@@ -39,6 +40,7 @@ describe('transition', () => {
 
         await new Promise(resolve => {
             setTimeout(() => {
+                expect(defaultXProps.onShow).toHaveBeenCalledTimes(1);
                 expect(result.current.status).toBe(STATUS.OPENING);
 
                 setTimeout(() => {