+++ webpack.config.js
@@ -8,8 +8,6 @@ module.exports = (env = {}) => {
     const MESSAGES_CONFIG = getWebpackConfig({
         entry: './src/index.js',
         filename: 'messaging.js',
-        modulename: ['paypal', 'Messages'],
-        libraryTarget: env.demo ? 'umd' : 'window',
         web: true,
         minify: true,
         debug: false,
@@ -20,7 +18,6 @@ module.exports = (env = {}) => {
             TARGET: 'standalone'
         })
     });
-    MESSAGES_CONFIG.output.libraryExport = 'Messages';
 
     // zoid components
     const COMPONENTS_CONFIG = getWebpackConfig({